Replaced explicit checks against true with Then { saved }
Prior to rspec-given-3.4.0, a bare failure object in a Then clause
would not re-raise an error.  Because a failure object was not false
or nil, the test would silently pass (which is NOT good).  Therefoce I
always used Then { saved == true }, which worked, but rather
overspecified the passing condition.

Version 3.4.0 of rspec-given added to_bool support to the failure
object, so that a bare failure will trigger the re-raising of the
exception.  Hence the removal of the "== true" checks.